# ⚡️ Speed up function `find_column_dependencies` by 233%